prediction set float64

+++ src/time_series_api.py
@@ -56,7 +56,7 @@ class TimeSeriesAPI:
         assert self.pred_calls == self.next_calls - 1, "You must get the next batch before making a new prediction."
         assert pred_df.columns.to_list() == ["row_id", "target"], "Prediction dataframe have invalid columns."
 
-        pred_df = pred_df.astype({"row_id": np.dtype("str"), "target": np.dtype("float32")})
+        pred_df = pred_df.astype({"row_id": np.dtype("str"), "target": np.dtype("float64")})
 
         self.predictions.append(pred_df)
         self.pred_calls += 1
